LayoutTests/imported/w3c/web-platform-tests/dom/nodes/moveBefore/moveBefore-selector-matching-expected.txt

This text should be green

FAIL moveBefore() should invalidate target when descendant selector changes assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
